improve docs and resolve review comments
