+++ src/lib.rs
@@ -361,30 +361,4 @@ mod tests {
         let loader: Box<dyn Loader> = Box::new(loader);
         assert_eq!("Hello World!", loader.lookup(&US_ENGLISH, "hello-world"));
     }
-
-    #[test]
-    fn check_if_multiloader_works() {
-        const US_ENGLISH: LanguageIdentifier = langid!("en-US");
-        const CHINESE: LanguageIdentifier = langid!("zh-CN");
-
-        let en_loader = ArcLoader::builder("./tests/locales", US_ENGLISH)
-            .customize(|bundle| bundle.set_use_isolating(false))
-            .build()
-            .unwrap();
-        let cn_loader = ArcLoader::builder("./tests/locales", CHINESE)
-            .customize(|bundle| bundle.set_use_isolating(false))
-            .build()
-            .unwrap();
-
-        let multiloader = MultiLoader::from_iter([
-            Box::new(en_loader) as Box<dyn Loader>,
-            Box::new(cn_loader) as Box<dyn Loader>,
-        ]);
-
-        assert_eq!(
-            "Hello World!",
-            multiloader.lookup(&US_ENGLISH, "hello-world")
-        );
-        assert_eq!("儿", multiloader.lookup(&CHINESE, "exists"));
-    }
 }
